[refactor] shadcn의 유틸리티 함수 경로 및 이름 수정
- utils.ts 이름을 cn.ts로 변경
  - 파일명을 기능에 맞게 명확히 수정
- 파일 위치를 FSD 방법론에 맞게 재배치
  - 이전: @/shared/lib/utils.ts
  - 이후: @/shared/lib/utils/cn.ts
