+++ src/shared/shadcnComponent/ui/input.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import {cn} from '@/shared/lib/utils';
+import {cn} from '@/shared/lib/utils/cn';
 
 const Input = React.forwardRef<HTMLInputElement, React.ComponentProps<'input'>>(({className, type, ...props}, ref) => {
   return (
